refactor(ui,theme): update preference component styles using yumaGlassCard
