+++ app/src/main/kotlin/moe/rukamori/archivetune/ui/screens/settings/BackupAndRestore.kt
@@ -301,19 +301,6 @@ fun BackupAndRestore(
                     )
                 }
             }
-
-            PreferenceGroup(title = stringResource(R.string.external_service)) {
-                spotifyAccountPreferences(
-                    state = spotifyState,
-                    showPlaylists = showSpotifyPlaylists,
-                    onConnectClick = { showSpotifyLogin = true },
-                    onShowPlaylistsChange = onShowSpotifyPlaylistsChange,
-                    onReloadClick = spotifyAccountViewModel::reloadPlaylists,
-                    onLogoutClick = {
-                        spotifyAccountViewModel.logout()
-                    },
-                )
-            }
         }
     }
 
@@ -419,490 +406,6 @@ fun BackupAndRestore(
         stepText = backupRestoreProgress?.step ?: progressStatus,
         indeterminate = backupRestoreProgress?.indeterminate ?: false,
     )
-}
-
-private fun PreferenceGroupScope.spotifyAccountPreferences(
-    state: SpotifyAccountUiState,
-    showPlaylists: Boolean,
-    onConnectClick: () -> Unit,
-    onShowPlaylistsChange: (Boolean) -> Unit,
-    onReloadClick: () -> Unit,
-    onLogoutClick: () -> Unit,
-) {
-    if (!state.isAuthenticated) {
-        item {
-            PreferenceEntry(
-                title = { Text(stringResource(R.string.spotify_connect)) },
-                description = stringResource(R.string.spotify_not_connected),
-                icon = { Icon(painterResource(R.drawable.spotify_icon), null) },
-                trailingContent = {
-                    AnimatedVisibility(visible = state.isLoading) {
-                        CircularWavyProgressIndicator(
-                            modifier = Modifier.size(28.dp),
-                            color = MaterialTheme.colorScheme.primary,
-                        )
-                    }
-                },
-                onClick = onConnectClick,
-                isEnabled = !state.isLoading,
-            )
-        }
-        return
-    }
-
-    item {
-        PreferenceEntry(
-            title = {
-                Text(
-                    text =
-                        if (state.accountName.isNotBlank()) {
-                            stringResource(R.string.spotify_connected_as, state.accountName)
-                        } else {
-                            stringResource(R.string.spotify_account)
-                        },
-                    maxLines = 1,
-                    overflow = TextOverflow.Ellipsis,
-                )
-            },
-            description =
-                when {
-                    state.isLoading -> stringResource(R.string.spotify_loading_library)
-                    state.playlistCount > 0 -> stringResource(R.string.spotify_available_count, state.playlistCount)
-                    else -> stringResource(R.string.spotify_no_sources)
-                },
-            icon = { SpotifyAccountIcon(avatarUrl = state.accountAvatarUrl) },
-            trailingContent = {
-                AnimatedVisibility(visible = state.isLoading) {
-                    CircularWavyProgressIndicator(
-                        modifier = Modifier.size(28.dp),
-                        color = MaterialTheme.colorScheme.primary,
-                    )
-                }
-            },
-            isEnabled = false,
-        )
-    }
-
-    item {
-        SwitchPreference(
-            title = { Text(stringResource(R.string.spotify_show_playlist)) },
-            description = stringResource(R.string.spotify_show_playlist_desc),
-            icon = { Icon(painterResource(R.drawable.spotify_icon), null) },
-            checked = showPlaylists,
-            onCheckedChange = onShowPlaylistsChange,
-            isEnabled = !state.isLoading,
-        )
-    }
-
-    item {
-        PreferenceEntry(
-            title = { Text(stringResource(R.string.spotify_reload_playlist)) },
-            description = stringResource(R.string.spotify_reload_playlist_desc),
-            icon = { Icon(painterResource(R.drawable.sync), null) },
-            onClick = onReloadClick,
-            isEnabled = !state.isLoading,
-        )
-    }
-
-    item {
-        PreferenceEntry(
-            title = { Text(stringResource(R.string.action_logout)) },
-            icon = { Icon(painterResource(R.drawable.logout), null) },
-            onClick = onLogoutClick,
-            isEnabled = !state.isLoading,
-        )
-    }
-}
-
-@Composable
-private fun SpotifyAccountIcon(avatarUrl: String?) {
-    val context = LocalContext.current
-    val requestSize = with(LocalDensity.current) { SpotifyAccountIconSize.roundToPx() }
-    val accountIcon = painterResource(R.drawable.spotify_icon)
-    val imageRequest =
-        remember(context, avatarUrl, requestSize) {
-            avatarUrl
-                ?.takeIf(String::isNotBlank)
-                ?.let {
-                    ImageRequest
-                        .Builder(context)
-                        .data(it)
-                        .size(requestSize)
-                        .build()
-                }
-        }
-
-    Box(
-        modifier =
-            Modifier
-                .fillMaxSize()
-                .clip(CircleShape)
-                .background(MaterialTheme.colorScheme.primaryContainer),
-        contentAlignment = Alignment.Center,
-    ) {
-        if (imageRequest != null) {
-            AsyncImage(
-                model = imageRequest,
-                contentDescription = null,
-                placeholder = accountIcon,
-                error = accountIcon,
-                contentScale = ContentScale.Crop,
-                modifier = Modifier.fillMaxSize(),
-            )
-        } else {
-            Icon(
-                painter = accountIcon,
-                contentDescription = null,
-                tint = MaterialTheme.colorScheme.onPrimaryContainer,
-                modifier = Modifier.size(24.dp),
-            )
-        }
-    }
-}
-
-@SuppressLint("SetJavaScriptEnabled")
-@Composable
-private fun SpotifyLoginSheet(
-    onDismiss: () -> Unit,
-    onCookiesCaptured: (spDc: String, spKey: String) -> Unit,
-) {
-    val sheetState = rememberModalBottomSheetState(skipPartiallyExpanded = true)
-    var webView by remember { mutableStateOf<WebView?>(null) }
-    var mainWebView by remember { mutableStateOf<WebView?>(null) }
-    var captured by remember { mutableStateOf(false) }
-
-    DisposableEffect(Unit) {
-        onDispose {
-            webView?.destroySpotifyLoginWebView()
-            mainWebView?.takeIf { it !== webView }?.destroySpotifyLoginWebView()
-            webView = null
-            mainWebView = null
-        }
-    }
-
-    BackHandler(enabled = webView != null) {
-        val activeWebView = webView
-        val rootWebView = mainWebView
-        when {
-            activeWebView?.canGoBack() == true -> {
-                activeWebView.goBack()
-            }
-
-            activeWebView != null && rootWebView != null && activeWebView !== rootWebView -> {
-                activeWebView.destroySpotifyLoginWebView()
-                webView = rootWebView
-            }
-
-            else -> {
-                onDismiss()
-            }
-        }
-    }
-
-    ModalBottomSheet(
-        modifier = Modifier.fillMaxHeight(),
-        onDismissRequest = onDismiss,
-        sheetState = sheetState,
-        shape = RoundedCornerShape(topStart = 28.dp, topEnd = 28.dp),
-        containerColor = MaterialTheme.colorScheme.surface,
-    ) {
-        Column(
-            modifier =
-                Modifier
-                    .fillMaxWidth()
-                    .fillMaxHeight()
-                    .padding(horizontal = 20.dp)
-                    .padding(bottom = 20.dp),
-            verticalArrangement = Arrangement.spacedBy(12.dp),
-        ) {
-            Text(
-                text = stringResource(R.string.spotify_login_title),
-                style = MaterialTheme.typography.titleLarge,
-                fontWeight = FontWeight.Bold,
-            )
-            Text(
-                text = stringResource(R.string.spotify_waiting_for_login),
-                style = MaterialTheme.typography.bodyMedium,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-            )
-            AndroidView(
-                modifier =
-                    Modifier
-                        .fillMaxWidth()
-                        .weight(1f)
-                        .clip(MaterialTheme.shapes.large),
-                factory = { context ->
-                    val container = FrameLayout(context)
-                    val spotifyWebView =
-                        WebView(context).apply {
-                            val cookieManager = CookieManager.getInstance()
-                            cookieManager.setAcceptCookie(true)
-                            cookieManager.setAcceptThirdPartyCookies(this, true)
-                            configureSpotifyLoginWebView()
-
-                            fun captureCookies(url: String?): Boolean {
-                                if (captured) return true
-                                val cookies = readSpotifyCookies(cookieManager, url)
-                                val spDc = cookies["sp_dc"].orEmpty()
-                                if (spDc.isBlank()) return false
-                                captured = true
-                                cookieManager.flush()
-                                onCookiesCaptured(spDc, cookies["sp_key"].orEmpty())
-                                return true
-                            }
-
-                            webViewClient =
-                                object : WebViewClient() {
-                                    override fun shouldOverrideUrlLoading(
-                                        view: WebView,
-                                        request: WebResourceRequest,
-                                    ): Boolean =
-                                        shouldOverrideSpotifyLoginUrl(
-                                            view = view,
-                                            url = request.url?.toString(),
-                                            captureCookies = { url -> captureCookies(url) },
-                                        )
-
-                                    @Deprecated("Deprecated in Java")
-                                    override fun shouldOverrideUrlLoading(
-                                        view: WebView,
-                                        url: String?,
-                                    ): Boolean =
-                                        shouldOverrideSpotifyLoginUrl(
-                                            view = view,
-                                            url = url,
-                                            captureCookies = { targetUrl -> captureCookies(targetUrl) },
-                                        )
-
-                                    override fun onPageStarted(
-                                        view: WebView,
-                                        url: String?,
-                                        favicon: android.graphics.Bitmap?,
-                                    ) {
-                                        captureCookies(url)
-                                    }
-
-                                    override fun onPageFinished(
-                                        view: WebView,
-                                        url: String?,
-                                    ) {
-                                        captureCookies(url)
-                                    }
-                                }
-                            webChromeClient =
-                                SpotifyLoginWebChromeClient(
-                                    container = container,
-                                    parentWebView = this,
-                                    captureCookies = { url -> captureCookies(url) },
-                                    onActiveWebViewChanged = { activeWebView -> webView = activeWebView },
-                                )
-                            webView = this
-                            mainWebView = this
-                            resetAuthWebViewSession(context, this) {
-                                loadUrl(SpotifyAuth.LOGIN_URL)
-                            }
-                        }
-                    container.addView(
-                        spotifyWebView,
-                        FrameLayout.LayoutParams(
-                            ViewGroup.LayoutParams.MATCH_PARENT,
-                            ViewGroup.LayoutParams.MATCH_PARENT,
-                        ),
-                    )
-                    container
-                },
-                update = {
-                    webView = webView ?: mainWebView
-                },
-            )
-        }
-    }
-}
-
-private fun WebView.destroySpotifyLoginWebView() {
-    stopLoading()
-    loadUrl("about:blank")
-    (parent as? ViewGroup)?.removeView(this)
-    destroy()
-}
-
-@SuppressLint("SetJavaScriptEnabled")
-private fun WebView.configureSpotifyLoginWebView() {
-    settings.apply {
-        javaScriptEnabled = true
-        domStorageEnabled = true
-        javaScriptCanOpenWindowsAutomatically = true
-        setSupportMultipleWindows(true)
-        setSupportZoom(true)
-        builtInZoomControls = true
-        displayZoomControls = false
-        mixedContentMode = WebSettings.MIXED_CONTENT_COMPATIBILITY_MODE
-        userAgentString = SpotifyLoginUserAgent
-    }
-}
-
-private class SpotifyLoginWebChromeClient(
-    private val container: FrameLayout,
-    private val parentWebView: WebView,
-    private val captureCookies: (String?) -> Boolean,
-    private val onActiveWebViewChanged: (WebView) -> Unit,
-) : WebChromeClient() {
-    override fun onCreateWindow(
-        view: WebView,
-        isDialog: Boolean,
-        isUserGesture: Boolean,
-        resultMsg: Message,
-    ): Boolean {
-        closePopupWebViews()
-
-        val popupWebView =
-            WebView(view.context).apply {
-                val cookieManager = CookieManager.getInstance()
-                cookieManager.setAcceptCookie(true)
-                cookieManager.setAcceptThirdPartyCookies(this, true)
-                configureSpotifyLoginWebView()
-                webViewClient =
-                    object : WebViewClient() {
-                        override fun shouldOverrideUrlLoading(
-                            view: WebView,
-                            request: WebResourceRequest,
-                        ): Boolean =
-                            shouldOverrideSpotifyLoginUrl(
-                                view = view,
-                                url = request.url?.toString(),
-                                captureCookies = captureCookies,
-                            )
-
-                        @Deprecated("Deprecated in Java")
-                        override fun shouldOverrideUrlLoading(
-                            view: WebView,
-                            url: String?,
-                        ): Boolean =
-                            shouldOverrideSpotifyLoginUrl(
-                                view = view,
-                                url = url,
-                                captureCookies = captureCookies,
-                            )
-
-                        override fun onPageStarted(
-                            view: WebView,
-                            url: String?,
-                            favicon: android.graphics.Bitmap?,
-                        ) {
-                            captureCookies(url)
-                        }
-
-                        override fun onPageFinished(
-                            view: WebView,
-                            url: String?,
-                        ) {
-                            captureCookies(url)
-                        }
-                    }
-            }
-
-        val transport = resultMsg.obj as? WebView.WebViewTransport ?: return false
-        container.addView(
-            popupWebView,
-            FrameLayout.LayoutParams(
-                ViewGroup.LayoutParams.MATCH_PARENT,
-                ViewGroup.LayoutParams.MATCH_PARENT,
-            ),
-        )
-        popupWebView.bringToFront()
-        popupWebView.requestFocus()
-        onActiveWebViewChanged(popupWebView)
-        transport.webView = popupWebView
-        resultMsg.sendToTarget()
-        return true
-    }
-
-    override fun onCloseWindow(window: WebView) {
-        window.destroySpotifyLoginWebView()
-        onActiveWebViewChanged(parentWebView)
-    }
-
-    private fun closePopupWebViews() {
-        for (index in container.childCount - 1 downTo 0) {
-            val child = container.getChildAt(index) as? WebView ?: continue
-            if (child !== parentWebView) {
-                child.destroySpotifyLoginWebView()
-            }
-        }
-        onActiveWebViewChanged(parentWebView)
-    }
-}
-
-private fun shouldOverrideSpotifyLoginUrl(
-    view: WebView,
-    url: String?,
-    captureCookies: (String?) -> Boolean,
-): Boolean {
-    if (captureCookies(url)) return true
-
-    val targetUrl = url?.takeIf(String::isNotBlank) ?: return false
-    if (targetUrl.isWebViewLoadableUrl()) return false
-
-    targetUrl.intentBrowserFallbackUrl()?.let { fallbackUrl -> view.loadUrl(fallbackUrl) }
-    return true
-}
-
-private fun String.isWebViewLoadableUrl(): Boolean {
-    val scheme = runCatching { Uri.parse(this).scheme?.lowercase() }.getOrNull()
-    return scheme == "http" ||
-        scheme == "https" ||
-        scheme == "javascript" ||
-        scheme == "data" ||
-        scheme == "blob"
-}
-
-private fun String.intentBrowserFallbackUrl(): String? =
-    runCatching { Intent.parseUri(this, Intent.URI_INTENT_SCHEME) }
-        .getOrNull()
-        ?.getStringExtra("browser_fallback_url")
-        ?.takeIf { it.isWebViewLoadableUrl() }
-
-private fun readSpotifyCookies(
-    cookieManager: CookieManager,
-    currentUrl: String?,
-): Map<String, String> {
-    val urls =
-        linkedSetOf(
-            "https://open.spotify.com",
-            "https://accounts.spotify.com",
-            "https://spotify.com",
-        )
-    currentUrl?.toSpotifyCookieOrigin()?.let(urls::add)
-    val cookies = linkedMapOf<String, String>()
-    cookieManager.flush()
-    urls.forEach { url ->
-        cookieManager
-            .getCookie(url)
-            ?.split(";")
-            ?.map(String::trim)
-            ?.filter(String::isNotBlank)
-            ?.forEach { part ->
-                val separator = part.indexOf('=')
-                if (separator <= 0) return@forEach
-                val key = part.substring(0, separator).trim()
-                val value = part.substring(separator + 1).trim()
-                if (key.isNotBlank()) {
-                    cookies[key] = value
-                }
-            }
-    }
-    return cookies
-}
-
-private fun String.toSpotifyCookieOrigin(): String? {
-    val uri = runCatching { Uri.parse(this) }.getOrNull() ?: return null
-    val host = uri.host?.lowercase() ?: return null
-    if (host != "spotify.com" && !host.endsWith(".spotify.com")) return null
-    val scheme =
-        uri.scheme
-            ?.takeIf { it.equals("https", ignoreCase = true) || it.equals("http", ignoreCase = true) }
-            ?: "https"
-    return "$scheme://$host"
 }
 
 @Composable
