+++ docs/design/YDS_ru.md
@@ -44,19 +44,16 @@ Yuma — это дизайн-система, ориентированная на
 ---
 
 ## 4. Философия компонентов
-
-* **Card:** Существует **только** для логического объединения нескольких связанных сущностей или настроек. Не является декоративной подложкой для одиночного текста.
-* **Button:** Кнопка всегда означает явное действие. Запрещено использовать кнопку в качестве контейнера для контента.
-* **Glass Surface:** Применяется **только** тогда, когда элемент визуально оторван от основного слоя контента и парит над ним.
-* **Accent Color:** Используется **только** для привлечения внимания пользователя к главному действию. Никогда не окрашивает фоновые контейнеры целиком.
-* **Divider:** Не используется. Границы контейнеров и отступы полностью закрывают задачу разделения.
+* **Card:** Используется для логической группировки настроек, деталей трека или метаданных. Сохраняет непрозрачный тоновый фон для максимального контраста и читаемости.
+* **Button & Chip:** Представляет явные действия или теги (например, аудиокодек, битрейт). Использует мягкие тоновые заливки (`surfaceContainerHigh`) с четкой типографикой.
+* **Overlay / Bottom Sheet:** Использует сплошные высококонтрастные фоновые поверхности (`surfaceContainerLow` / `surfaceContainer`) с мягкими скругленными углами. Размещается поверх затемненного блюра заднего плана (`BackdropBlur`), создавая глубину без потери читаемости контента.
+* **Accent Color:** Зарезервирован исключительно для активных состояний, переключателей и главных действий. Избегает заливки крупных фоновых поверхностей.
+* **Divider:** Не используется. Отступы в макете (16–24dp) и границы контейнеров обеспечивают естественное разделение.
 
 ---
 
 ## 5. Токены (Foundations)
-
 ### Отступы (Spacing)
-
 * Micro — 4dp
 * Small — 8dp
 * Medium — 16dp
@@ -64,32 +61,15 @@ Yuma — это дизайн-система, ориентированная на
 * Extra Large — 32dp
 
 ### Радиусы скругления (Radius)
-
 * Small — 12dp
-* Medium — 16dp
-* Large — 24dp
-* Max — 32dp / Pill / Circle
+* Medium — 16dp (Кнопки, Чипы, Элементы списка)
+* Large — 24dp (Диалоги, Bottom Sheets, Карточки медиа)
+* Max — 32dp / Pill / Circle (FAB, Плавающий плеер, Переключатели, Контейнеры иконок)
 
-### Цвета (Colors)
-
-* **Primary, Secondary, Tertiary** — динамически генерируются из палитры текущего контента.
-* **SurfaceContainer** — базовые сплошные поверхности.
-* **GlassBackground** (Alpha 0.65f) + **GlassBorder** (Alpha 0.25f) — поверхностные оверлеи.
-
-### Типографика (Typography)
-
-* **Display / Headline** — акцентные крупные заголовки.
-* **Title / Body** — системная читаемая типографика.
-
-### Движение (Motion)
-
-* Пружинные анимации (`Spring`) с акцентом на естественную физическую отдачу.
-
-### Прозрачность (Opacity)
-
-* Disabled — 0.38f
-* Secondary Text — 0.65f
-* Pressed Overlay — 0.12f
+### Цвета и поверхности (Colors & Surfaces)
+* **Primary, Secondary, Tertiary** — динамические цвета M3 Expressive, генерируемые из текущей обложки.
+* **Base Surfaces** — `surface`, `surfaceContainer`, `surfaceContainerHigh` (непрозрачные тоновые слои для штор, карточек и диалогов).
+* **Backdrop Blur** — применяется строго к фоновому слою за плавающими оверлеями/диалогами. Без внутренних стеклянных рамок или прозрачных текстовых карточек.
 
 ---
 
@@ -155,20 +135,17 @@ Yuma — это дизайн-система, ориентированная на
 ---
 
 ## 10. Do / Don't
-
 ### ✔ DO (Делай)
-
-* Использовать отступы вместо разделителей.
-* Объединять связанные настройки и контент в крупные контейнеры.
-* Использовать один динамический акцентный цвет для главных действий.
-* Оставлять стеклянные поверхности только для плавающего UI (мини-плеер, диалоги, BottomSheet, меню).
+* Использовать сплошные тоновые контейнеры (`surfaceContainer`) внутри диалогов и Bottom Sheets для максимальной читаемости.
+* Применять размытие фона (Backdrop Blur) позади плавающих оверлеев для создания визуальной иерархии.
+* Использовать единые круглые контейнеры (`CircleShape`) для иконок-бейджей в настройках.
+* Сохранять четкие отступы в макете (16–24dp) вместо разделительных линий.
 
 ### ✘ DON'T (Не делай)
-
-* Не вкладывать карточку в карточку без явной логической причины.
-* Не смешивать стеклянные и сплошные поверхности внутри одной группы контента.
-* Не использовать более двух разных радиусов скругления в рамках одного экрана.
-* Не делать интерактивные элементы меньше **48×48dp**.
+* Использовать полупрозрачные стеклянные карточки с внутренними рамками для плотного текста или технической информации.
+* Смешивать более двух размеров радиусов скругления внутри одной группы компонентов.
+* Создавать интерактивные элементы управления меньше 48×48dp.
+* Перегружать крупные фоновые контейнеры динамическими акцентными цветами.
 
 ---
